Proof-of-concept for using react-hotkeys
Typing command+up while in the log-categories editor will take you
straight to the home page, and command+down will take you to the About
page.

So far, so good. Unfortunately, being wrapped in a <HotKeys> element
seems to affect the rendering of some child-components, in a way that
I have not yet got to the bottom of. Maybe some bad interaction
between react-hotkeys and react-bootstrap?

Part of STRIPES-13.