stripes-components dep consistent with stripes-core
The stripes-components dependency was out of sync with stripes-core,
which was causing funky overlaps of the settings pane where the third
column would be superimposed on the second. This probably means we have
our Pane and Paneset components in the wrong repository, or maybe that
we are managing that dependency incorrectly. In any case, the version of
stripes-components we get from stripes-core needs to be the same as the
one we depend on here. And now it is.